--- cs/TagCloud/ITagCloud.cs
@@ -1,11 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagCloud
-{
-    public interface ITagCloud
-    {
-        void PutNextTag(Size tagSize);
-
-        string SaveBitmap(bool shouldShowLayout, bool shouldShowMarkup, bool openAfterSave);
-    }
-}
