--- cs/TagCloud/Layouting/ICloudLayouter.cs
@@ -1,11 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagCloud.Layouting
-{
-    public interface ICloudLayouter : ILayouter
-    {
-        Point Center { get; }
-
-        int GetCloudBoundaryRadius();
-    }
-}
