--- cs/TagCloud/Layouting/ILayouter.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Drawing;
-
-namespace TagCloud.Layouting
-{
-    public interface ILayouter
-    {
-        Size GetRectanglesBoundaryBox();
-
-        List<Rectangle> GetRectanglesCopy();
-
-        Rectangle PutNextRectangle(Size rectangleSize);
-    }
-}
