--- cs/TagCloud/Saving/IBitmapSaver.cs
@@ -1,9 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagCloud.Saving
-{
-    public interface IBitmapSaver
-    {
-        string Save(Bitmap bitmap, bool openAfterSave);
-    }
-}
