+++ cs/TagCloudTask/TagCloudTask.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31729.503
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloudTask", "TagCloudTask\TagCloudTask.csproj", "{5AD022EA-367E-43AC-B455-6C5440555455}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestDataGenerator", "TestDataGenerator\TestDataGenerator.csproj", "{D2C3C561-7727-4E68-B107-EC2B585E7BC5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloudTaskTests", "TagCloudTaskTests\TagCloudTaskTests.csproj", "{3337306F-91A7-45C7-90C1-4A25EA726A09}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5AD022EA-367E-43AC-B455-6C5440555455}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5AD022EA-367E-43AC-B455-6C5440555455}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5AD022EA-367E-43AC-B455-6C5440555455}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5AD022EA-367E-43AC-B455-6C5440555455}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D2C3C561-7727-4E68-B107-EC2B585E7BC5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D2C3C561-7727-4E68-B107-EC2B585E7BC5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D2C3C561-7727-4E68-B107-EC2B585E7BC5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D2C3C561-7727-4E68-B107-EC2B585E7BC5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3337306F-91A7-45C7-90C1-4A25EA726A09}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3337306F-91A7-45C7-90C1-4A25EA726A09}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3337306F-91A7-45C7-90C1-4A25EA726A09}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3337306F-91A7-45C7-90C1-4A25EA726A09}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {55363CF4-C800-411A-B278-A9577BAB9A7F}
+	EndGlobalSection
+EndGlobal
