+++ cs/TagCloudTask/TagCloudTask/ITagCloud.cs
@@ -1,0 +1,11 @@
+﻿using System.Drawing;
+
+namespace TagCloudTask
+{
+    public interface ITagCloud
+    {
+        void PutNextTag(Size tagSize);
+
+        string SaveBitmap(bool shouldShowLayout, bool shouldShowMarkup, bool openAfterSave);
+    }
+}
