+++ cs/TagCloudTask/TagCloudTask/Layouting/ICloudLayouter.cs
@@ -1,0 +1,11 @@
+﻿using System.Drawing;
+
+namespace TagCloudTask.Layouting
+{
+    public interface ICloudLayouter : ILayouter
+    {
+        Point Center { get; }
+
+        int GetCloudBoundaryRadius();
+    }
+}
