+++ cs/TagCloudTask/TagCloudTask/Saving/IBitmapSaver.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagCloudTask.Saving
+{
+    public interface IBitmapSaver
+    {
+        string Save(Bitmap bitmap, bool openAfterSave);
+    }
+}
