+++ cs/tdd.sln
@@ -7,12 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BowlingGame", "BowlingGame\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Samples", "Samples\Samples.csproj", "{B5108E20-2ACF-4ED9-84FE-2A718050FC94}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloudVisualization", "TagCloud\TagCloudVisualization.csproj", "{5CBE717A-D750-45BE-B9E0-46CB36DDEA00}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TagCloudVisualization_Tests", "TagCloud_Tests\TagCloudVisualization_Tests.csproj", "{137617C9-58FD-4615-930B-5C32BA3C8AB7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloud_TestDataGenerator", "TagCloud_TestDataGenerator\TagCloud_TestDataGenerator.csproj", "{259037C3-4FA5-439F-B92E-3CB09A1DEF33}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,18 +21,6 @@ Global
 		{B5108E20-2ACF-4ED9-84FE-2A718050FC94}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B5108E20-2ACF-4ED9-84FE-2A718050FC94}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B5108E20-2ACF-4ED9-84FE-2A718050FC94}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5CBE717A-D750-45BE-B9E0-46CB36DDEA00}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5CBE717A-D750-45BE-B9E0-46CB36DDEA00}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5CBE717A-D750-45BE-B9E0-46CB36DDEA00}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5CBE717A-D750-45BE-B9E0-46CB36DDEA00}.Release|Any CPU.Build.0 = Release|Any CPU
-		{137617C9-58FD-4615-930B-5C32BA3C8AB7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{137617C9-58FD-4615-930B-5C32BA3C8AB7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{137617C9-58FD-4615-930B-5C32BA3C8AB7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{137617C9-58FD-4615-930B-5C32BA3C8AB7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{259037C3-4FA5-439F-B92E-3CB09A1DEF33}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{259037C3-4FA5-439F-B92E-3CB09A1DEF33}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{259037C3-4FA5-439F-B92E-3CB09A1DEF33}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{259037C3-4FA5-439F-B92E-3CB09A1DEF33}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
